move webrtc tests in common pkg
